Enforce order between the cron job and mirror file
If you setup a server using this class just before 4am UTC, its possible
for the cron job to be written out before the mirror list is written.
When this happens, the default mirror list file provided by the
apt-mirror package is used. This ends up mirroring the Ubuntu archive.
This bug was discovered when a newly built node had a full HDD.
Enforcing the ordering should ensure that the mirror list is written
first so that you do not unintentionally mirror any packages.